Parameters order with hash_equals
There're 2 parameters in built-in function hash_equals, the first is
"known_string" and second is "user_string".

The notes in PHP manual of hash_equals:
> It is important to provide the user-supplied string as the second
> parameter, rather than the first.

Otherwise, there's a question in stackoverflow.com about [Why is order
of arguments in PHP's hash_equals() function important](https://stackoverflow.com/questions/27911597/why-is-order-of-arguments-in-phps-hash-equals-function-important)

In this commit, it changed paramenter orders for hash_equals, becuase
the "user_string" is provided by `$this->route('id')` and
`$this->route('hash')`.